-Finished conversion. Very efficient and parallel computation now. When loading 1K images crashes at imageclass building though...

+++ include/Util.hpp
@@ -8,7 +8,7 @@ class Util {
   public:
     static void generate_random_string(char *s, const int len);
     //static float euclideanDistanceImageActivations(const map<string,CNNLayer> &layers, const Image &img1, const Image &img2);
-    static float euclideanDistanceImageClass(const ImageClass &imgc1, const ImageClass &imgc2, const CNNScheme &scheme);
+    static double euclideanDistanceImageClass(const ImageClass &imgc1, const ImageClass &imgc2, const CNNScheme &scheme);
     static void computeImageClasses(vector<Image> &images, const CNNScheme &scheme,  vector<ImageClass> &imageClasses);
 }; // UTIL
